+++ Banner/RUNA/Banner/RUNABannerViewExtension.m
@@ -50,7 +50,7 @@
     }
 }
 
--(void) setLocationWithLatitude:(float) lat longitude:(float) lon {
+-(void) setLocationWithLatitude:(double) lat longitude:(double) lon {
     if (lat < -90.0 || lat > 90.0) {
         NSLog(@"[RUNA] illegal latitude value, must be from -90.0 to +90.0, where nagative is south!");
         return;
